Complete 04_removeFromArrray
Annoyed about this one as I had to look at the solution. Have been
trying not to do this but it was so complicated and made no sense to
me at all. Everything I googled for help didn't provide great help.

I think I understand it all now but a very complicated one.